+++ services/dbox.py
@@ -14,7 +14,7 @@ class File(Document): # Now I don't know how it shoud work
 
 
 
-base64_DROPBOX_ACCESS_TOKEN = 'c2wuQlJsNDY3ZWQza19CU1ZxX0xCQk1YOVkzRHJ6V3NnLWpuWmRud2lLN2FVcFlMT0hKdl9nZ2oya05iaG1SUVA2cGY0dUFWNXdFSTNSdDYtcUdOM2Qtd29CWmo0R0RCYXZweWtyXzZOWE81SXRHZG5qSVRzM0pqeGRGYkdtbHFJejQwY3UyRy1iMTlBeGw='
+base64_DROPBOX_ACCESS_TOKEN = 'c2wuQlJ6VXJvV1p1NU9NcWpDSzVXWVpSWWxmejRkWkdVOFc1aEVIOHhtbHdUREYzajY5MVdaR2s3eHlLS0I1SDBYcVRJeHNIYTh0UkVvWml0NlU4d2FQbWJYUFVsWWFaOVJWSk0yT0M2VVRtRmdqa0p0dE9xcWNyakNzSzhrYXl0UEwtYkFjWmJaLW80THQ='
 
 
 base64_bytes = base64_DROPBOX_ACCESS_TOKEN.encode('ascii')
