+++ templates/components/delete-modal.html
@@ -19,7 +19,7 @@
         ></button>
       </div>
       <div class="modal-body">
-        <form action="/api/v1/web/contacts" method="POST">
+        <form action="/contacts" method="POST">
           <!-- Name input -->
           <div class="form-outline mb-4">
             <input
