+++ templates/components/edit-modal.html
@@ -19,7 +19,7 @@
         ></button>
       </div>
       <div class="modal-body">
-        <form action="/api/v1/web/contacts" method="POST">
+        <form action="/contacts" method="POST">
           <!-- Hidden -->
           <input type="hidden" name="new_contact-id" value="this.{{contact.id}}">
           <!-- Name input -->
