+++ templates/files/files.html
@@ -75,7 +75,7 @@
     {% endif %}
     {% if token %}
     <section>
-      <form method="post" action="/api/v1/web/uploadfiles" enctype="multipart/form-data" class="d-flex align-content-center justify-content-center gap-3">
+      <form method="post" action="/uploadfiles" enctype="multipart/form-data" class="d-flex align-content-center justify-content-center gap-3">
         <div class="w-50">
           <label for="file-text" class="form-label">Upload your common Files</label>
           <input class="form-control" name="file" id="file" type="file" accept="image/*,.pdf,.doc," size="" onchange="document.getElementById('file-text').textContent = this.files.length > 0 ? this.files[0].name : 'Select a file'">
@@ -97,7 +97,7 @@
     <div class="w-100">
     Your own dropbox cloud
         {% if user_token %}
-        <form method="post" action="/api/v1/web/uploadfiles" enctype="multipart/form-data" class="d-flex align-content-center justify-content-center gap-3">
+        <form method="post" action="/uploadfiles" enctype="multipart/form-data" class="d-flex align-content-center justify-content-center gap-3">
           <div class="w-50">
             <label for="file-text" class="form-label">Upload your own Files</label>
             <input class="form-control" name="user_file" id="user_file" type="file" accept="image/*,.pdf,.doc," size="" onchange="document.getElementById('file-text').textContent = this.files.length > 0 ? this.files[0].name : 'Select a file'">
@@ -122,7 +122,7 @@
           {%endif%}
         {% endif %}
         {% if not user_token %}
-        <form action="/api/v1/web/uploadfiles" method="post">
+        <form action="/uploadfiles" method="post">
           <input type="text" id="token" name="token" class="token" value="Enter access token"/>
           <button>Get access!</button>
         </form>
