+++ templates/index.html
@@ -48,13 +48,13 @@
           <div class="d-flex gap-4 mt-5">
             <a
               class="btn btn-primary btn-dark btn-lg"
-              href="/api/v1/web/signup"
+              href="/signup"
               role="button"
               >Signup</a
             >
             <a
               class="btn btn-primary btn-light btn-lg"
-              href="/api/v1/web/presentation"
+              href="/presentation"
               role="button"
               >Presentation</a
             >
@@ -63,7 +63,7 @@
           <div class="d-flex gap-4 mt-5">
             <a
               class="btn btn-primary btn-light btn-lg"
-              href="/api/v1/web/presentation"
+              href="/presentation"
               role="button"
               >Presentation</a
             >
