+++ templates/login.html
@@ -49,7 +49,7 @@
           <div
             class="col-md-9 col-lg-6 col-xl-5 d-flex justify-content-center align-items-center"
           >
-            <a href="/api/v1/web/" class="back-to-home">
+            <a href="/" class="back-to-home">
               <img
                 src="{{ url_for('static', path='/image/big-owl.svg') }}"
                 class="img-fluid"
@@ -100,7 +100,7 @@
                 role="tabpanel"
                 aria-labelledby="tab-login"
               >
-                <form action="/api/v1/web/signup" method="POST">
+                <form action="/signup" method="POST">
                   <!-- Email input -->
                   <div class="form-outline mb-4">
                     <input
@@ -137,7 +137,7 @@
                 role="tabpanel"
                 aria-labelledby="tab-register"
               >
-                <form action="/api/v1/web/signup" method="POST">
+                <form action="/signup" method="POST">
                   <!-- Name input -->
                   <div class="form-outline mb-4">
                     <input
