+++ templates/notes/notes_dashboard.html
@@ -72,7 +72,7 @@ crossorigin="anonymous"
                     <div class="d-flex align-items-center">
                         <!-- <span class="mr-1"><i class="fa fa-star favourite-note"></i></span>
                         <span class="mr-1"> -->
-                        <form action="/api/v1/web/notes" method="POST">
+                        <form action="/notes" method="POST">
                             <input type="hidden" name="note-id" id="note-id" value="{{note.id}}">
                          <button
                          
@@ -119,7 +119,7 @@ crossorigin="anonymous"
     </div>
 
     <!-- Modal Add notes -->
-    <form  action="/api/v1/web/notes" method="POST">
+    <form  action="/notes" method="POST">
     <div class="modal fade" id="addnotesmodal" tabindex="-1" role="dialog" aria-labelledby="addnotesmodalTitle" style="display: none;" aria-hidden="true">
         <div class="modal-dialog modal-dialog-centered" role="document">
             <div class="modal-content border-0">
